feat(config): support tools.codex.fast_mode for persistent fast mode (#1429)
Add fast_mode: Option<bool> to both ToolConfig (project config) and
GlobalToolConfig (global config), allowing users to set fast mode once
in config rather than passing --fast-but-more-cost on every invocation.

Merge logic: CLI flag OR project config OR global config. Any true value
enables codex fast mode. The run/review/debate paths all apply this
precedence consistently. Warning functions updated to fire on effective
fast mode (config or CLI).

Extracted RunLoopRequest/RunLoopCompletion/RunLoopOutcome types to
run_cmd_attempt_types.rs and build_debate_instruction to
debate_cmd_readonly.rs to stay under 800-line monolith limit.
