+++ crates/cli-sub-agent/src/run_cmd_attempt.rs
@@ -1,15 +1,12 @@
 use anyhow::Result;
 use csa_config::{ExecutionEnvOptions, GlobalConfig, ProjectConfig};
-use csa_core::types::{OutputFormat, ToolName, ToolSelectionStrategy};
+use csa_core::types::{OutputFormat, ToolSelectionStrategy};
 use csa_executor::structured_output_instructions_for_fork_call;
 use csa_lock::slot::{
     SlotAcquireResult, ToolSlot, acquire_slot_blocking, format_slot_diagnostic, slot_usage,
     try_acquire_slot,
 };
-use std::{
-    path::{Path, PathBuf},
-    time::Instant,
-};
+use std::time::Instant;
 use tracing::{info, warn};
 
 use super::attempt_exec::{
@@ -39,68 +36,9 @@ use crate::run_cmd_tool_selection::{
 };
 use crate::run_helpers::{is_tool_binary_available_for_config, parse_tool_name};
 
-pub(crate) struct RunLoopRequest<'a> {
-    pub(crate) strategy: ToolSelectionStrategy,
-    pub(crate) initial_tool: ToolName,
-    pub(crate) initial_model_spec: Option<String>,
-    pub(crate) user_model_spec_explicit: bool,
-    pub(crate) initial_model: Option<String>,
-    pub(crate) runtime_fallback_candidates: Vec<ToolName>,
-    pub(crate) project_root: &'a Path,
-    pub(crate) config: Option<&'a ProjectConfig>,
-    pub(crate) global_config: &'a GlobalConfig,
-    pub(crate) prompt_text: &'a str,
-    pub(crate) skill: Option<&'a str>,
-    pub(crate) skill_session_tag: Option<String>,
-    pub(crate) description: Option<String>,
-    pub(crate) parent: Option<String>,
-    pub(crate) output_format: OutputFormat,
-    pub(crate) stream_mode: csa_process::StreamMode,
-    pub(crate) thinking: Option<&'a str>,
-    pub(crate) force: bool,
-    pub(crate) force_override_user_config: bool,
-    pub(crate) force_ignore_tier_setting: bool,
-    pub(crate) no_failover: bool,
-    pub(crate) fast_but_more_cost: bool,
-    pub(crate) wait: bool,
-    pub(crate) idle_timeout_seconds: u64,
-    pub(crate) cli_idle_timeout: Option<u64>,
-    pub(crate) cli_initial_response_timeout: Option<u64>,
-    pub(crate) no_idle_timeout: bool,
-    pub(crate) run_timeout_seconds: Option<u64>,
-    pub(crate) run_started_at: Instant,
-    pub(crate) is_fork: bool,
-    pub(crate) is_auto_seed_fork: bool,
-    pub(crate) ephemeral: bool,
-    pub(crate) fork_call: bool,
-    pub(crate) session_arg: Option<String>,
-    pub(crate) effective_session_arg: Option<String>,
-    pub(crate) tier_auto_select: bool,
-    pub(crate) failover_on_crash_enabled: bool,
-    pub(crate) resolved_tier_name: Option<&'a str>,
-    pub(crate) context_load_options: Option<&'a csa_executor::ContextLoadOptions>,
-    pub(crate) memory_injection: pipeline::MemoryInjectionOptions,
-    pub(crate) pre_session_hook: Option<csa_hooks::PreSessionHookInvocation>,
-    pub(crate) task_needs_edit: Option<bool>,
-    pub(crate) no_fs_sandbox: bool,
-    pub(crate) extra_writable: Vec<PathBuf>,
-    pub(crate) extra_readable: Vec<PathBuf>,
-    pub(crate) branch_guard: crate::run_helpers_branch_guard::BranchGuardRuntime,
-}
-
-pub(crate) enum RunLoopCompletion {
-    Exit(i32),
-    Completed(Box<RunLoopOutcome>),
-}
-
-pub(crate) struct RunLoopOutcome {
-    pub(crate) result: csa_process::ExecutionResult,
-    pub(crate) current_tool: ToolName,
-    pub(crate) executed_session_id: Option<String>,
-    pub(crate) fork_resolution: Option<ForkResolution>,
-    /// Ordered list of tools skipped due to rate-limit/quota failover before this result.
-    pub(crate) fallback_chain: csa_scheduler::FallbackChain,
-}
+#[path = "run_cmd_attempt_types.rs"]
+mod types;
+pub(crate) use types::{RunLoopCompletion, RunLoopOutcome, RunLoopRequest};
 
 pub(crate) async fn execute_run_loop(request: RunLoopRequest<'_>) -> Result<RunLoopCompletion> {
     // Compute max failover attempts: count total models across ALL tiers to
@@ -170,7 +108,19 @@ pub(crate) async fn execute_run_loop(request: RunLoopRequest<'_>) -> Result<RunL
             matches!(request.strategy, ToolSelectionStrategy::Explicit(_)),
         )
         .await?;
-        if request.fast_but_more_cost {
+        let fast_mode = request.fast_but_more_cost
+            || request
+                .config
+                .and_then(|c| c.tools.get("codex"))
+                .and_then(|t| t.fast_mode)
+                .unwrap_or(false)
+            || request
+                .global_config
+                .tools
+                .get("codex")
+                .and_then(|t| t.fast_mode)
+                .unwrap_or(false);
+        if fast_mode {
             executor.enable_codex_fast_mode();
         }
